refactor(server): controller, service 위치 및 네이밍 변경
