+++ server/src/purchase/purchase.controller.ts
@@ -5,10 +5,14 @@ import { PurchaseService } from './purchase.service';
 import {
   Purchase,
   purchaseSchema,
+  productSchema,
   CATEGORY_VALUES,
   PurchaseDetail,
+  ProductDetail,
   GetPurchaseDetailResponse,
+  GetProductDetailResponse,
   EditPurchaseResponse,
+  EditPurchaseItemResponse,
 } from '@jumble/shared';
 
 export const PurchaseController = {
@@ -101,62 +105,6 @@ export const PurchaseController = {
     }
   },
 
-  // 사용 안 되는 듯?
-  getPurchases: async (req: Request, res: Response) => {
-    try {
-      const userId = req.user.id;
-      const page = parseInt(req.query.page as string) || 1;
-      const limit = parseInt(req.query.limit as string) || 10;
-
-      if (page < 1 || limit < 1) {
-        return res.status(400).json({
-          success: false,
-          status: 400,
-          message: 'page와 limit은 1 이상이어야 합니다.',
-        });
-      }
-
-      const { purchases, total } = await PurchaseService.getPurchases(userId, page, limit);
-
-      return res.status(200).json({
-        success: true,
-        status: 200,
-        message: '사입내역 조회에 성공했습니다.',
-        purchases,
-        pagination: {
-          total,
-          page,
-          limit,
-          totalPages: Math.ceil(total / limit),
-        },
-      });
-    } catch (error) {
-      console.error('🚨 서버 에러 발생:', error);
-
-      if (error instanceof Prisma.PrismaClientKnownRequestError) {
-        return res.status(400).json({
-          success: false,
-          status: 400,
-          message: '잘못된 요청입니다.',
-        });
-      }
-
-      if (error instanceof Prisma.PrismaClientValidationError) {
-        return res.status(400).json({
-          success: false,
-          status: 400,
-          message: '요청 데이터 형식이 올바르지 않습니다.',
-        });
-      }
-
-      return res.status(500).json({
-        success: false,
-        status: 500,
-        message: '서버 오류가 발생했습니다.',
-      });
-    }
-  },
-
   // 사입내역 상세조회 API
   getPurchase: async (req: Request, res: Response) => {
     try {
@@ -188,6 +136,56 @@ export const PurchaseController = {
         message: '사입내역 상세 조회에 성공했습니다.',
         data: purchase,
       } satisfies GetPurchaseDetailResponse);
+    } catch (error) {
+      console.error('🚨 서버 에러 발생:', error);
+
+      if (error instanceof Prisma.PrismaClientKnownRequestError) {
+        return res.status(400).json({
+          success: false,
+          status: 400,
+          message: '잘못된 요청입니다.',
+        });
+      }
+
+      return res.status(500).json({
+        success: false,
+        status: 500,
+        message: '서버 오류가 발생했습니다.',
+      });
+    }
+  },
+
+  // 상품사입내역 상세조회 API
+  getProduct: async (req: Request, res: Response) => {
+    try {
+      const userId = req.user.id;
+      const rawItemId = req.params.productId as string;
+
+      if (!/^\d+$/.test(rawItemId)) {
+        return res.status(400).json({
+          success: false,
+          status: 400,
+          message: 'id는 정수여야 합니다.',
+        });
+      }
+
+      const itemId = BigInt(rawItemId);
+      const item: ProductDetail | null = await PurchaseService.getProduct(userId, itemId);
+
+      if (!item) {
+        return res.status(404).json({
+          success: false,
+          status: 404,
+          message: '상품 사입내역을 찾을 수 없습니다.',
+        });
+      }
+
+      return res.status(200).json({
+        success: true,
+        status: 200,
+        message: '상품 사입내역 상세 조회에 성공했습니다.',
+        data: item,
+      } satisfies GetProductDetailResponse);
     } catch (error) {
       console.error('🚨 서버 에러 발생:', error);
 
@@ -257,6 +255,72 @@ export const PurchaseController = {
             message: '유효하지 않은 참조 값입니다.',
           });
         }
+      }
+
+      return res.status(500).json({
+        success: false,
+        status: 500,
+        message: '서버 오류가 발생했습니다.',
+      });
+    }
+  },
+
+  // 상품사입내역 수정 API
+  updateProduct: async (req: Request, res: Response) => {
+    try {
+      const userId = req.user.id;
+      const rawItemId = req.params.productId as string;
+
+      if (!/^\d+$/.test(rawItemId)) {
+        return res.status(400).json({
+          success: false,
+          status: 400,
+          message: 'id는 정수여야 합니다.',
+        });
+      }
+
+      const result = productSchema.safeParse(req.body);
+      if (!result.success) {
+        return res.status(400).json({
+          success: false,
+          status: 400,
+          message: result.error.issues[0].message,
+        });
+      }
+
+      const itemId = BigInt(rawItemId);
+      const found = await PurchaseService.updateProduct(userId, itemId, result.data);
+
+      if (!found) {
+        return res.status(404).json({
+          success: false,
+          status: 404,
+          message: '상품 사입내역을 찾을 수 없습니다.',
+        });
+      }
+
+      return res.status(200).json({
+        success: true,
+        status: 200,
+        message: '상품 사입내역 수정에 성공했습니다.',
+      } satisfies EditPurchaseItemResponse);
+    } catch (error) {
+      console.error('🚨 서버 에러 발생:', error);
+
+      if (error instanceof Prisma.PrismaClientKnownRequestError) {
+        return res.status(400).json({
+          success: false,
+          status: 400,
+          message: '잘못된 요청입니다.',
+        });
+      }
+
+      if (error instanceof Prisma.PrismaClientValidationError) {
+        return res.status(400).json({
+          success: false,
+          status: 400,
+          message: '요청 데이터 형식이 올바르지 않습니다.',
+        });
       }
 
       return res.status(500).json({
