+++ server/src/purchase/purchase.service.ts
@@ -1,7 +1,7 @@
-import { Purchase, Product } from '@jumble/shared';
+import { Purchase, Product, ProductDetail } from '@jumble/shared';
 import prisma from '../lib/prisma';
 import { serializeBigInt } from '../utils/serializeBigInt';
-import { formatPurchase } from '../utils/format';
+import { formatPurchase, formatPurchaseItem } from '../utils/format';
 
 export const PurchaseService = {
   // 사입내역 추가 API
@@ -59,47 +59,6 @@ export const PurchaseService = {
     });
   },
 
-  // 사용 안 되는 듯?
-  getPurchases: async (userId: bigint, page: number, limit: number) => {
-    const skip = (page - 1) * limit;
-    const where = { user_id: userId };
-
-    const [purchases, total] = await prisma.$transaction([
-      prisma.purchase.findMany({
-        where,
-        skip,
-        take: limit,
-        orderBy: { purchased_at: 'desc' },
-        select: {
-          id: true,
-          purchase_no: true,
-          purchased_at: true,
-          vendor: { select: { name: true } },
-          items: {
-            select: {
-              id: true,
-              purchase_item_no: true,
-              item_name: true,
-              category: true,
-              color: true,
-              size: true,
-              extra_option: true,
-              unit_price: true,
-              quantity: true,
-              backorder_quantity: true,
-            },
-          },
-          receipt: { select: { receipt_image_url: true } },
-        },
-      }),
-      prisma.purchase.count({ where }),
-    ]);
-
-    const formattedPurchases = purchases.map(formatPurchase);
-
-    return { purchases: serializeBigInt(formattedPurchases), total };
-  },
-
   // 사입내역 상세조회 API
   getPurchase: async (userId: bigint, purchaseId: bigint) => {
     const purchase = await prisma.purchase.findFirst({
@@ -130,6 +89,46 @@ export const PurchaseService = {
     if (!purchase) return null;
 
     return serializeBigInt(formatPurchase(purchase));
+  },
+
+  // 상품사입내역 상세조회 API
+  getProduct: async (userId: bigint, itemId: bigint): Promise<ProductDetail | null> => {
+    const item = await prisma.purchaseItem.findFirst({
+      where: { id: itemId, purchase: { user_id: userId } },
+      select: {
+        id: true,
+        purchase_item_no: true,
+        item_name: true,
+        category: true,
+        color: true,
+        size: true,
+        extra_option: true,
+        unit_price: true,
+        quantity: true,
+        backorder_quantity: true,
+        purchase: {
+          select: {
+            id: true,
+            purchase_no: true,
+            purchased_at: true,
+            vendor: { select: { name: true } },
+          },
+        },
+      },
+    });
+
+    if (!item) return null;
+
+    const { purchase, ...rest } = item;
+    const formatted = {
+      purchaseId: purchase.id,
+      purchaseNo: purchase.purchase_no,
+      purchasedAt: purchase.purchased_at,
+      vendor: purchase.vendor.name,
+      ...formatPurchaseItem(rest),
+    };
+
+    return serializeBigInt(formatted);
   },
 
   // 사입내역 수정 API
@@ -180,6 +179,31 @@ export const PurchaseService = {
 
       return true;
     });
+  },
+
+  // 상품사입내역 수정 API
+  updateProduct: async (userId: bigint, itemId: bigint, data: Product): Promise<boolean | null> => {
+    const existing = await prisma.purchaseItem.findFirst({
+      where: { id: itemId, purchase: { user_id: userId } },
+    });
+
+    if (!existing) return null;
+
+    await prisma.purchaseItem.update({
+      where: { id: itemId },
+      data: {
+        item_name: data.name,
+        category: data.category,
+        color: data.color,
+        size: data.size,
+        extra_option: data.option,
+        unit_price: data.price,
+        quantity: data.quantity,
+        backorder_quantity: data.backorderQuantity,
+      },
+    });
+
+    return true;
   },
 
   getPurchaseReceipt: async (userId: bigint, purchaseId: bigint) => {
